requirements.txt
Django==1.10.6
python3-saml==1.2.3
psycopg2==2.7
django-webpack-loader==0.4.1
django-oauth-toolkit==0.12.0
markdown==2.6.8
djangorestframework==3.6.2
django_filter==1.0.1
Pillow==4.0.0
